Make debugSamplingClock protected and refactor tests.
