+++ packages/flutter/test/gestures/gesture_binding_resample_event_on_widget_test.dart
@@ -16,18 +16,25 @@ import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+class TestResampleEventFlutterBinding extends AutomatedTestWidgetsFlutterBinding {
+  @override
+  SamplingClock? get debugSamplingClock => TestSamplingClock(this.clock);
+}
+
 class TestSamplingClock implements SamplingClock {
   TestSamplingClock(this._clock);
 
   @override
   DateTime now() => _clock.now();
-  @override Stopwatch stopwatch() => _clock.stopwatch();
+
+  @override
+  Stopwatch stopwatch() => _clock.stopwatch();
 
   final Clock _clock;
 }
 
 void main() {
-  final TestWidgetsFlutterBinding binding = AutomatedTestWidgetsFlutterBinding();
+  final TestWidgetsFlutterBinding binding = TestResampleEventFlutterBinding();
   testWidgets('PointerEvent resampling on a widget', (WidgetTester tester) async {
     assert(WidgetsBinding.instance == binding);
     Duration currentTestFrameTime() => Duration(milliseconds: binding.clock.now().millisecondsSinceEpoch);
@@ -94,7 +101,6 @@ void main() {
     GestureBinding.instance!.resamplingEnabled = true;
     const Duration kSamplingOffset = Duration(milliseconds: -5);
     GestureBinding.instance!.samplingOffset = kSamplingOffset;
-    GestureBinding.instance!.debugSamplingClock = TestSamplingClock(binding.clock);
     ui.window.onPointerDataPacket!(packet);
     expect(events.length, 0);
